+++ lib/ModelMixins/SearchableItemMixin.ts
@@ -1,13 +1,15 @@
 import { action, computed } from "mobx";
 import Constructor from "../Core/Constructor";
-import { ItemSearchProviders } from "../Models/ItemSearchProviders";
 import ItemSearchProvider, {
   ItemSearchResult
 } from "../Models/ItemSearchProvider";
+import { ItemSearchProviders } from "../Models/ItemSearchProviders";
 import Model from "../Models/Model";
 import SearchableItemTraits from "../Traits/SearchableItemTraits";
 
 type MixinModel = Model<SearchableItemTraits>;
+
+export type ItemSelectionDisposer = () => void;
 
 /**
  * This mixin adds capability for searching a catalog item using an {@link
@@ -18,23 +20,26 @@ function SearchableItemMixin<T extends Constructor<MixinModel>>(Base: T) {
     readonly hasSearchableItemMixin = true;
 
     /**
-     * Callback when a search result is selected by the user.
+     * A hook for highlighting search results selected by the user.
      *
      * The implementation can decide how to highlight the search result.
      *
-     * @param result The selected search result.
+     * @param results The search results to be highlighted.
      */
-    abstract selectItemSearchResult(result: ItemSearchResult): void;
+    abstract highlightItemSearchResults(
+      results: ItemSearchResult[]
+    ): ItemSelectionDisposer;
 
     /**
-     * Callback when user un-selects a search result.
+     * A hook for hiding search results selected by the user.
      *
-     * This provides a chance to reverse the highlighting and changes made
-     * during the call to {@selectItemSearchResult}
+     * The implementation can decide how to highlight the search result.
      *
-     * @param result The un-selected search result.
+     * @param results The search results to be hidden.
      */
-    abstract unselectItemSearchResult(result: ItemSearchResult): void;
+    abstract hideItemSearchResults(
+      results: ItemSearchResult[]
+    ): ItemSelectionDisposer;
 
     /**
      * An optional implementation for zooming in to results.